+++ F0V1/EntryPoint_AS.cpp
@@ -112,12 +112,6 @@ namespace
                 .function("bool Update()", &System::Update);
         }
 
-        // asbind20::value_class<KeyboardInput>(engine, "KeyboardInput", asOBJ_POD | asOBJ_APP_CLASS_ALLINTS)
-        //     .behaviours_by_traits()
-        //     .method("bool down() const", &KeyboardInput::down)
-        //     .method("bool pressed() const", &KeyboardInput::pressed)
-        //     .method("bool up() const", &KeyboardInput::up);
-
         KeyboardInput::RegisterScript(engine);
 
         for (auto handler : g_asapi_globalBindHandlers)
